adding WASHES2026 as main
