+++ src/pages/WASHES/index.jsx
@@ -1,0 +1,3 @@
+export { default as Washes2024 } from "./Washes2024";
+export { default as Washes2025 } from "./Washes2025";
+export { default as Washes2026 } from "./Washes2026";
